Adding profile to command line parser (fix?)
While trying out the command with the `--profile` flag with traces on,
I've noticed the profile wasn't passed through the AWS call.

`AWS_PROFILE` could override it, and looking at the code, it looks like
an omission (region was there twice).

Feel free to discard this PR if this was intentional.